requirements.txt
# Core dependencies
torch==2.2.2
numpy>=1.24.0
gymnasium>=1.0.0

# Visualization and logging
matplotlib>=3.7.0
imageio==2.35.1
imageio[ffmpeg]
wandb>=0.16.0
pandas>=2.0.0
tensorboard>=2.12.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0